Fix tests
Importing main.js causes the router to be initialized which kicks off some
initial actions. The tests don't expect this and failed. Moving routes to their
own module let's us import them directly for testing.
